company DAO added up

+++ src/main/java/app/entities/Logs.java
@@ -6,11 +6,11 @@ public class Logs {
 
     private final Employee employee;
     private final Project project;
-    private final int time;
+    private final double time;
     private final String comment;
     private final Date date;
 
-    public Logs(Employee employee, Project project, int time, String comment, Date date) {
+    public Logs(Employee employee, Project project, double time, String comment, Date date) {
         this.employee = employee;
         this.project = project;
         this.time = time;
